% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactacao.r
\name{acumulacens}
\alias{acumulacens}
\title{Compactacao Por Valor Acumulado}
\usage{
acumulacens(cenarios, quebras = 1L)
}
\arguments{
\item{cenarios}{objeto da classe \code{cenarios} contendo apenas uma bacia e ano de referencia}

\item{quebras}{ou um inteiro indicando em quantas partes iguais separar o dado ou um vetor de
inteiros indicando as posicoes nas quais separar}
}
\value{
objeto da classe \code{compactcen} contendo o dado em dimensao reduzida
}
\description{
Reducao de dimensao por valor acumulada, possivelmente em partes crescentes
}
\details{
Esta funcao reduz a dimensao de um vetor de cenarios calculando a soma acumulada do vetor,
possivelmente em partes. Deve ser notado que, antes da compactacao, os cenarios sao escalonados
para o intervalo \eqn{[0, 1]} de modo a obter valores de menor variancia. Esta regularizacao e feita
por bacia individualmente. Isto objetiva determinar um espaco reduzido que capture principalmente
o perfil de cada cenario em cada regiao, independentemente da magnitude.

Se \code{quebras = 3L}, por exemplo, a funcao quebra o cenario em tres partes iguais e calcula a
soma de cada um. O vetor em dimensao reduzida e, entao, a soma acumulada destas partes. Para
cenarios de um ano, indo de janeiro a dezembro, o uso de \code{quebras = 3L} representa o calculo
do acumulado ate abril, agosto e dezembro (por construcao um vetor de valores crescentes).
}
