% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-cluster.r
\name{addnewobs}
\alias{addnewobs}
\title{Adicao De Novos Dados}
\usage{
addnewobs(clust, newcompact)
}
\arguments{
\item{clust}{objeto contendo a clusterizacao a qual adicionar novas obs}

\item{newcompact}{objeto \code{compactcen} contendo novas observacoes ja compactadas. Idealmente,
esta compactacao deve ser feita pela mesma funcao utilizada nos dados originais}
}
\value{
objeto \code{clust} alterado com novas observacoes e suas classes ao final dos dados
originais
}
\description{
Generica para adicao de novas observacoes a clusters ja estimados
}
\details{
Cada metodos age sobre o objeto \code{clust} o minimo possivel para que as novas observacoes
sejam vistas pelos metodos \code{getclustclass} e \code{getclustmeans}. Isto significa que
diversos atributos do objeto \code{clust} nao sao alterados e entao se tornam incoerentes - ex.
a variacao total entre e intra clusters do kmeans.
}
