% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-cenarios.r
\name{as.cenarios}
\alias{as.cenarios}
\title{Conversao Para \code{cenarios}}
\usage{
as.cenarios(dat)
}
\arguments{
\item{dat}{\code{data.frame} ou \code{data.table} para converter}
}
\value{
objeto da classe \code{cenarios} -- veja \code{\link{new_cenarios}} para mais detalhes
}
\description{
Converte um \code{data.frame} ou \code{data.table} em um objeto da classe \code{cenarios}
}
\details{
\code{dat} deve ter uma estrutura especifica, de quatro conlunas nomeadas
\code{c("grupo", "cenario", "indice", "valor")}, nao necessariamente nesta ordem.

\code{grupo} pode ser uma coluna de inteiros, fatores, strings ou qualquer outro tipo de dado,
indicando de alguma forma grupos aos quais os cenarios pertencem como, por exemplo, usinas ou
bacias. \code{cenario} deve ser uma coluna de inteiros indicando o indice do cenario.
\code{indice} deve ser uma coluna de inteiros, datas ou qualquer outro tipo de dado, indicando de
alguma forma a sequencia temporal dos valores no cenario. Por fim, \code{valor} deve ser uma
coluna numerica indicando o valor simulado.

Um exemplo de \code{dat} e dado abaixo\tabular{cccc}{
   grupo \tab cenario \tab indice \tab valor \cr
   SIN \tab 1 \tab 2022-01-01 \tab XXXX \cr
   SIN \tab 1 \tab 2022-01-02 \tab XXXX \cr
   SIN \tab 1 \tab 2022-01-03 \tab XXXX \cr
   SIN \tab 2 \tab 2022-01-01 \tab XXXX \cr
   SIN \tab 2 \tab 2022-01-02 \tab XXXX \cr
}
}
\examples{

# usando um dado inventado
dat <- data.frame(
    grupo = rep(paste0("G", seq(3)), each = 30),
    cenario = rep(rep(seq(6), each = 5), 3),
    indice = rep(seq(5), 18),
    valor = rnorm(90)
 )

cens <- as.cenarios(dat)
\dontrun{
cens
}

# caso o dado nao possua as colunas nomeadas como esperado, serao adicionadas na ordem que se
# esperava encontrar e um aviso sera emitido
dat2 <- data.frame(
    rep(paste0("G", seq(3)), each = 30),
    rep(rep(seq(6), each = 5), 3),
    rep(seq(5), 18),
    rnorm(90)
 )
\dontrun{
cens2 <- as.cenarios(dat2) # assume que as colunas estao na ordem esperada
cens2
}

# por fim, se o dado tiver um numero de colunas diferente de quatro, a coercao retorna erro
dat3 <- data.frame(
    rep(paste0("G", seq(3)), each = 30),
    rep(rep(seq(6), each = 5), 3),
    rep(seq(5), 18),
    rnorm(90),
    rep(NA, 90)
 )
\dontrun{
cens3 <- as.cenarios(dat3)
}

}
\seealso{
\code{\link{[.cenarios}} para subset de objetos da classe \code{cenarios}
}
