% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustcens.r
\docType{package}
\name{clustcens}
\alias{clustcens}
\title{Pacote \code{clustcens}}
\description{
Selecao E Extracao De Cenarios
}
\details{
Este pacote prove funcoes para selecao de cenarios de variaveis genericas, independentemente da
metodologia utilizada para gera-los. Alem dos metodos para selecao de cenarios sao fornecidas
ferramentas para visualizacao da escolha.
}
