% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-cluster.r
\name{clustkmeans}
\alias{clustkmeans}
\alias{getclustclass.kmeans}
\alias{getclustmeans.kmeans}
\alias{addnewobs.kmeans}
\alias{addnewobs.Mclust}
\title{Clusteriza Dado Por Kmeans}
\usage{
clustkmeans(compact, nc, nstart = 30, ...)

\method{getclustclass}{kmeans}(clust)

\method{getclustmeans}{kmeans}(clust)

\method{addnewobs}{kmeans}(clust, newcompact)

\method{addnewobs}{Mclust}(clust, newcompact)
}
\arguments{
\item{compact}{objeto \code{compactcen} contendo cenarios compactados para clusterizar}

\item{nc}{numero de clusters}

\item{nstart}{numero de sementes para testar o kmeans}

\item{...}{demais parametros passados a funcao \code{\link[stats]{kmeans}} exceto \code{nstart}}

\item{clust}{objeto da classe \code{Mclust}}

\item{newcompact}{objeto \code{compactcen} contendo novas observacoes a adicionar aos clusters}
}
\value{
objeto \code{kmeans} -- Veja \code{\link[stats]{kmeans}} para mais detalhes
}
\description{
Wrapper da funcao \code{\link[stats]{kmeans}} para uso neste pacote
}
