% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-cenarios.r
\name{new_cenarios}
\alias{new_cenarios}
\title{Construtor Interno De \code{cenarios}}
\usage{
new_cenarios(dat)
}
\arguments{
\item{dat}{um data.table contendo as colunas \code{c("grupo", "cenario", "indice", "valor")}}
}
\value{
objeto \code{cenarios}, uma lista de um elemento chamado "cenarios" que e um
\code{data.table} com as colunas

\describe{
\item{\code{grupo}}{grupo ao qual o cenario corresponde}
\item{\code{cenario}}{indice do cenario com respeito a bacia e ano de referencia}
\item{\code{indice}}{indice sequencial de valores do cenario}
\item{\code{valor}}{valor do cenario}
}

Adicionalmente contem os atributos

\itemize{
\item{\code{grupos}: }{vetor de caracters contendo as grupos presentes no dado}
\item{\code{indices}: }{vetor contendo os indices para os quais foram gerados cenarios}
\item{\code{ncen}: }{inteiro indicando numero de cenarios por grupo}
}
}
\description{
Funcao interna para instanciar objetos \code{cenarios}, nao deve ser chamada diretamente
}
