% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact-PCA.r
\name{predict.PCAcens}
\alias{predict.PCAcens}
\title{Compactacao De Novas Observacoes}
\usage{
\method{predict}{PCAcens}(object, newcens, ...)
}
\arguments{
\item{object}{objeto \code{compactcen} contendo uma compactacao por PCA}

\item{newcens}{objeto \code{cenarios} contendo novos valores a compactar}

\item{...}{nao tem uso, existe apenas para consistencia com a generica}
}
\value{
objeto da classe \code{compactcen} contendo o novo dado em dimensao reduzida
}
\description{
Compacta novos cenarios utilizando a mesma funcao ja estimada para o objeto \code{x}
}
