% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-cenarios.r
\name{[.cenarios}
\alias{[.cenarios}
\title{Subset De \code{cenarios}}
\usage{
\method{[}{cenarios}(x, i, j, k, ...)
}
\arguments{
\item{x}{objeto da classe \code{cenarios}}

\item{i}{escalar ou vetor de grupos para extrair do dado}

\item{j}{escalar ou vetor de indices de cenarios para extrair do dado}

\item{k}{escalar ou vetor de datas para extrair do dado}

\item{...}{nao possui uso -- existe apenas para consistencia com a generica}
}
\value{
objeto \code{cenarios} contendo apenas os valores especificados
}
\description{
Extrai subsets de objetos \code{cenarios} e adequa os atributios
}
