% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact-PCA.r
\name{monta_out_PCA}
\alias{monta_out_PCA}
\title{Monta Objeto \code{compactcens} Com PCA}
\usage{
monta_out_PCA(compdat, SIGMA, importance, cenarios, grupos, escala)
}
\arguments{
\item{importance}{inteiro indicando quantas componentes reter}

\item{compact}{matriz contendo a transformacao completa do dado em PCAs}

\item{cens}{nomes dos cenarios originalmente passados no objeto sendo compactado}
}
\value{
objeto da classe \code{compactcen} contendo o dado em dimensao reduzida
}
\description{
Wrapper de codigo comum entre \code{PCAcens} e seu metodo \code{predict}
}
