% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactacao.r
\name{PCAcens}
\alias{PCAcens}
\title{Compactacao Por PCA}
\usage{
PCAcens(cenarios, vartot = 0.8)
}
\arguments{
\item{cenarios}{objeto da classe \code{cenarios} contendo apenas uma bacia e ano de referencia}

\item{vartot}{percentual em formato decimal de variacao total mininima}
}
\value{
objeto da classe \code{compactcen} contendo o dado em dimensao reduzida
}
\description{
Reduz dimensionalidade dos cenarios via PCA mantendo um minimo de variacao total
}
\details{
A matriz de dados considerada para reducao de dimensionalidade tem em cada linha o cenario e,
em cada coluna, os passos a frente de simulacao por bacia. Desta forma, cem cenarios 15 passos a
frente de dois grupos corresponde a uma matriz de dados 100 por 30. Serao selecionadas as n
primeiras componentes principais que representem no minimo \code{vartot} por cento da variacao
total. Deve ser observado que o dado e normalizado para media zero e variancia um antes da
compactacao.
}
