% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{plot.cenarios}
\alias{plot.cenarios}
\title{Visualizacao De \code{cenarios}}
\usage{
\method{plot}{cenarios}(x, ..., print = TRUE)
}
\arguments{
\item{x}{objeto da classe \code{cenarios}}

\item{...}{objetos \code{cenarios} opcionais com cenarios a serem plotados por cima}

\item{print}{booleano indicando se o plot deve ser exibido ou retornado invisivelmente}
}
\description{
Plot em grade de objetos \code{cenarios}
}
\details{
O argumento \code{...} permite que sejam passados mais outros objetos \code{cenarios} contendo
um numero menor de cenarios selecionados para plot por cima do completo.
}
